requirements.txt
python-telegram-bot==21.0.1
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
